Partial Web Version is now working
A number of changes has happened. Firstly a Plugins folder is added so javascript code can be run. This folder has a js file called serverFunction.js, inside it, is a browser based get request, however it is synchronous. Secondly GameController, LevelFactory, OldMenu has been changed accordingly so that it would work on web or editor or standalone build.