Fixed unnecessary call to the server
In GameController, it would call to the server each time a new game happens. However when a new game happens, the Cinematic will always be first before the game starts. The Cinematic is the one that should be calling to the server and storing the necessary info in the GlobalState.